+++ processing/topics/parking/off_street_parking/helper/result_tags_off_street_parking.lua
@@ -28,7 +28,7 @@ local function result_tags_off_street_parking(result, area)
 
   local result_meta = Metadata(result)
 
-  local cleaned_tags, replaced_tags = sanitize_cleaner(result_tags, result.object.tags)
+  local cleaned_tags, replaced_tags = sanitize_cleaner.split_cleaned_and_replaced_tags(result_tags, result.object.tags)
 
   return {
     id = DefaultId(result.object),