+++ processing/topics/parking/roads/helper/result_tags_roads.lua
@@ -3,15 +3,15 @@ require("CopyTags")
 require("DefaultId")
 require("Metadata")
 require("RoadClassificationRoadValue")
-local is_driveway_helper = require('is_driveway')
-local has_parking = require('has_parking')
+local is_driveway_check = require('is_driveway')
+local has_parking_check = require('has_parking')
 local SANITIZE_TAGS = require('sanitize_tags')
 require("road_width")
 
 function result_tags_roads(object)
   local id = DefaultId(object)
   local width, width_confidence, width_source = road_width(object.tags)
-  local is_driveway = is_driveway_helper(object.tags)
+  local is_driveway = is_driveway_check(object.tags)
 
   local result_tags = {
     highway = object.tags.highway,
@@ -19,7 +19,7 @@ function result_tags_roads(object)
     name = SANITIZE_TAGS.road_name(object.tags),
     category = is_driveway and "driveway" or "road",
     is_driveway = is_driveway,
-    has_parking = has_parking(object.tags),
+    has_parking = has_parking_check(object.tags),
     width = width,
     width_confidence = width_confidence,
     width_source = width_source,