+++ processing/topics/roads_bikelanes/bikelanes/helper/deriveTrafficMode.lua
@@ -2,7 +2,7 @@ require('init')
 require('Set')
 local SANITIZE_ROAD_TAGS = require('sanitize_road_tags')
 local SANITIZE_PARKING_TAGS = require('sanitize_parking_tags')
-local SANITIZE_VALUES = require('sanitize_values')
+local sanitize_cleaner = require('sanitize_cleaner')
 
 local DIRECTIONAL_PARKING_INFERENCE_CATEGORIES = Set({
   'cyclewayOnHighway_advisory',
@@ -17,9 +17,8 @@ local function inferTrafficModeFromParking(centerlineTags, side)
   local raw_value = centerlineTags['parking:' .. side] or centerlineTags['parking:both']
   local value = SANITIZE_PARKING_TAGS.parking(raw_value)
 
-  if value == nil then return nil end
+  if sanitize_cleaner.remove_disallowed_value(value) == nil then return nil end
   if value == 'no' then return nil end
-  if value == SANITIZE_VALUES.disallowed then return nil end
   return 'parking'
 end