+++ processing/CHANGELOG.md
@@ -6,6 +6,8 @@ This is an manual and incomplete list of change to the data structure.
 
 ### `bikelanes`
 
+- Add attributes `operator_type`, `informal`, `covered`
+- Include bikelanes for ways that are privately operated, indoor or informal
 - Include bikelanes that are `access=destination` in the data
 - Categorize bicycle roads with "Kfz frei" as `bicycleRoad_vehicleDestination`